refactor: core cache design and impl
