+++ pkg/cache/trace_test.go
