+++ pkg/plugins/gateway/algorithms/least_kv_cache.go
@@ -38,11 +38,11 @@ func init() {
 }
 
 type leastKvCacheRouter struct {
-	cache *cache.Cache
+	cache cache.Cache
 }
 
 func NewLeastKvCacheRouter() (Router, error) {
-	c, err := cache.GetCache()
+	c, err := cache.Get()
 	if err != nil {
 		return nil, err
 	}
@@ -68,12 +68,12 @@ func (r leastKvCacheRouter) Route(ctx context.Context, pods map[string]*v1.Pod, 
 		// Due to metric refactor (pull/543) to better support lora and multi models,
 		// we change to use PodModelMetrics instead of PodMetrics in some scenarios.
 		// This works but doesn't look very promising, we can revisit this part later.
-		gpuCache, err := r.cache.GetPodModelMetric(pod.Name, routingCtx.Model, metrics.GPUCacheUsagePerc)
+		gpuCache, err := r.cache.GetMetricValueByPodModel(pod.Name, routingCtx.Model, metrics.GPUCacheUsagePerc)
 		if err != nil {
 			klog.Error(err)
 			continue
 		}
-		cpuCache, err := r.cache.GetPodModelMetric(pod.Name, routingCtx.Model, metrics.CPUCacheUsagePerc)
+		cpuCache, err := r.cache.GetMetricValueByPodModel(pod.Name, routingCtx.Model, metrics.CPUCacheUsagePerc)
 		if err != nil {
 			klog.Error(err)
 			continue
