+++ pkg/plugins/gateway/algorithms/least_request.go
@@ -39,11 +39,11 @@ func init() {
 }
 
 type leastRequestRouter struct {
-	cache *cache.Cache
+	cache cache.Cache
 }
 
 func NewLeastRequestRouter() (Router, error) {
-	c, err := cache.GetCache()
+	c, err := cache.Get()
 	if err != nil {
 		return nil, err
 	}
@@ -67,17 +67,17 @@ func (r leastRequestRouter) Route(ctx context.Context, pods map[string]*v1.Pod, 
 	}
 
 	for _, pod := range readyPods {
-		runningReq, err := r.cache.GetPodModelMetric(pod.Name, routingCtx.Model, metrics.NumRequestsRunning)
+		runningReq, err := r.cache.GetMetricValueByPodModel(pod.Name, routingCtx.Model, metrics.NumRequestsRunning)
 		if err != nil {
 			klog.Error(err)
 			continue
 		}
-		waitingReq, err := r.cache.GetPodModelMetric(pod.Name, routingCtx.Model, metrics.NumRequestsWaiting)
+		waitingReq, err := r.cache.GetMetricValueByPodModel(pod.Name, routingCtx.Model, metrics.NumRequestsWaiting)
 		if err != nil {
 			klog.Error(err)
 			continue
 		}
-		swappedReq, err := r.cache.GetPodModelMetric(pod.Name, routingCtx.Model, metrics.NumRequestsSwapped)
+		swappedReq, err := r.cache.GetMetricValueByPodModel(pod.Name, routingCtx.Model, metrics.NumRequestsSwapped)
 		if err != nil {
 			klog.Error(err)
 			continue
