+++ pkg/plugins/gateway/gateway_req_body.go
@@ -56,7 +56,7 @@ func (s *Server) HandleRequestBody(ctx context.Context, requestID string, req *e
 	}
 
 	// early reject the request if model doesn't exist.
-	if !s.cache.CheckModelExists(model) {
+	if !s.cache.GetModel(model) {
 		klog.ErrorS(nil, "model doesn't exist in cache, probably wrong model name", "requestID", requestID, "model", model)
 		return generateErrorResponse(envoyTypePb.StatusCode_BadRequest,
 			[]*configPb.HeaderValueOption{{Header: &configPb.HeaderValue{
@@ -65,7 +65,7 @@ func (s *Server) HandleRequestBody(ctx context.Context, requestID string, req *e
 	}
 
 	// early reject if no pods are ready to accept request for a model
-	pods, err := s.cache.GetPodsForModel(model)
+	pods, err := s.cache.ListPodsByModel(model)
 	if len(pods) == 0 || len(utils.FilterReadyPods(pods)) == 0 || err != nil {
 		klog.ErrorS(err, "no ready pod available", "requestID", requestID, "model", model)
 		return generateErrorResponse(envoyTypePb.StatusCode_ServiceUnavailable,
